perf: move /teams route to (main-nav) route group
